restructure test and add test send email, #412
