+++ lib/handlers/email/post_sendmail.js
@@ -9,7 +9,7 @@ const candidatesEmailId     = require('../../helpers/candidates_email_id');
 const stringToArray         = require('../../helpers/string_to_array');
 const generateSignature     = require('../../helpers/generate_signature.js');
 const capitalizeFirstLetter = require('../../helpers/capitalize_first_letter.js');
-const btoa                  = require('btoa'); // encode email string to base64
+const btoa                  = require('btoa');
 const updateEmailsCandidate = require('../../database-helpers/elasticsearch/update_emails_candidate');
 
 module.exports = function (request, reply) {
@@ -84,7 +84,7 @@ module.exports = function (request, reply) {
 
         let emailObject = {};
         emailObject.subject = subject;
-        emailObject.message = message;
+        emailObject.message = email.message.replace(/{name}/g, candidate.firstName);
         emailObject.senderName = email.senderName;
         emailObject.senderId = email.senderId;
         emailObject.timestamp = email.timestamp;
@@ -92,9 +92,6 @@ module.exports = function (request, reply) {
 
         updateEmailsCandidate(candidate.id, emailObject, function(errorUpdate, responseUpdate) {
 
-          console.log(errorUpdate);
-          console.log(responseUpdate);
-          console.log('message sent');
         })
 
       }
